requirements.txt
h5py==3.11.0
hydra-core==1.3.2
imageio==2.34.2
matplotlib==3.9.0
networkx==3.3
nibabel==5.2.1
numpy==1.17.4
omegaconf==2.3.0
open3d==0.18.0
pandas==2.2.2
Pillow==7.0.0
Pillow==10.3.0
pytorch_lightning==2.3.0
pyvista==0.43.10
rdkit==2024.3.1
scikit_learn==1.5.0
scipy==1.3.3
seaborn==0.13.2
setuptools==45.2.0
timm==1.0.7
torch==2.3.1
torch_geometric==2.5.3
torch_scatter==2.1.2
torchmetrics==1.4.0.post0
tqdm==4.66.4
wandb==0.17.3
